Linted Code. Restructured Directories
