+++ E1_ShuffleNet/requirements.txt
